pyproject.toml
[tool.poetry]
name = "backend-default"
version = "0.1.0"
description = "Backend default para proyectos orientados a microservicios"
authors = ["Oyhs-co <oyhsotelo@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "api"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
supabase = ">=2.15.2,<3.0.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
sqlalchemy = ">=2.0.41,<3.0.0"
python-multipart = ">=0.0.20,<0.0.21"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
alembic = ">=1.16.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pika = ">=1.3.2,<2.0.0"
httpx = ">=0.28.1,<0.29.0"
pydantic = {extras = ["email"], version = "^2.11.5"}
autoflake = "^2.3.1"
requests = "^2.32.3"
types-sqlalchemy = "^1.4.53.38"
typing-extensions = "^4.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.0"
pytest-cov = "^6.1.1"
requests = "^2.32.3"
pytest-asyncio = "^1.0.0"

[tool.poetry.scripts]
start = "uvicorn api.gateway.services.projects.app.main:app"
test = "pytest -v"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.13.3"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true